tests to decimal, added stored procedures to migration

+++ Server.Core/Dtos/GetRecipeByCategoryDto.cs
@@ -4,6 +4,6 @@
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public float TotalCost { get; set; }
+        public decimal TotalCost { get; set; }
     }
 }
